Create removeAll
2. Array List (Dynamic) - removeAll
The List ADT has the following operations already implemented in class:

void add(int num): Inserts num into the list.
int get(int pos): Returns the integer in the *pos*th position.
int remove(int num): Removes the first occurrence of num from the list and returns the position to which it was found.
 

Your task is to implement the following operations on a Dynamic Array List:

int removeAll(int num): Removes all occurrences of the element num. This returns the number of elements removed.
void dynamic_deduce(): This method shall be called to dynamically reduce the
 

Note that when the number of elements reach 3/4 of the capacity rounded down, you are to reallocate the array by reducing its size by 20% rounded down. Example, when the array has capacity 12 and after removing all the *num*s and it reaches less than or equal to 9 elements, it reduces the capacity by 20% (2.4, rounded down to 2). Therefore the new capacity shall be 10.

 

Note that this time, only do the reallocation once per method call. This is so that there will only be at most one call of the realloc for this method. Additionally, add the functionality to maintain a minimum capacity of 5.

 

For this problem, the dynamic_add and the remove method are given in order to guide you with the tasks above.