Circular Linked List
You are to design a structure known as a circularly linked list, which is essentially a singularly linked list in which the next reference of the tail node is set to refer back to the head of the list (rather than null).

 

The Circular Linked List only has a reference to its tail and the size of the list. Since it is circular, the head can be accessed from the tail's next. We keep the reference of its tail instead of its head since we can access the head from the tail, but not vice-versa since this is a singularly linked list and has no reference to its previous node.

 

The following methods must be supported by the Circular Linked List:

void addFirst(int)
This method will add the given integer as the first element. Refer to the image below for the guide on how to perform addFirst. You must achieve a running time of O(1) in this method.



void addLast(int)
This method will add the given integer to the last element. You must achieve a running time of O(1) in this method.

int getTail()
This method will simply return the element of the tail. If the list is empty, return -1. You must achieve a running time of O(1) in this method.

int removeFirst()
This method will return and remove the first element. If the list is empty, return -1. You must achieve a running time of O(1) in this method.

void rotate()
This method will rotate the circular linked list. Refer to the image below for the guide on how to perform this method. You must achieve a running time of O(1) in this method.



void print()
This method will print the size of the list, then onto a new line the first element until the last element (separated by ->) by the use of the rotate() method. At the end, it should also print the next element of the last element to illustrate a Circular Linked List. Of course, ensure that the tail must be the same before and after this method has been called. If the list is empty, output "Empty" instead of printing the elements.

 

For all the above methods, you may utilize existing methods for your implementation.