Doubly Linked List+ - removeFirst and removeLast
In class, we had a group activity that creates and utilizes the removeNode(node* n) method to remove an item. We have not used it in our removeFirst and removeLast yet.

 

You are to implement the removeNode method again.

Improve the current implementation of the removeFirst and removeLast to only have a few lines that calls the removeNode method. Note that when the list is empty, you are not to remove anything. You must not use the size to determine if the list is empty, instead use other ways to determine whether the list is empty. When you do use the size, you'll receive deductions for this problem.