Doubly Linked List - combine
Given a basic sentinel-less Doubly Linked List, implement the combine(DoublyLL*) method that will connect the contents of the DoublyLL* parameter to this instance's linked list.

 

For example, this linked list has 10 <-> 20 <-> 30 <-> 40 and the parameter DoublyLL has 70 <-> 80, this linked list should combine what is in the parameter to its own, resulting to this linked list having 10 <-> 20 <-> 30 <-> 40 <-> 70 <-> 80. The parameter DoublyLL should also be reset to not contain anything and size back to 0.

 

Note that this operation must be O(1) and you are not to create more nodes as you are only to connect the contents of the list.